+++ regions/io/ds9/write.py
@@ -1,15 +1,27 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
+import os
+
+from astropy.utils.decorators import deprecated
+
+from ...core import Region, Regions
+from ...core.registry import RegionsRegistry
 from ..core import _to_shape_list
 
 __all__ = ['write_ds9', 'ds9_objects_to_string']
 
 
+@RegionsRegistry.register(Region, 'serialize', 'ds9')
+@RegionsRegistry.register(Regions, 'serialize', 'ds9')
+def _serialize_ds9(regions, coordsys='fk5', fmt='.6f', radunit='deg'):
+    shapelist = _to_shape_list(regions, coordsys)
+    return shapelist.to_ds9(coordsys, fmt, radunit)
+
+
+@deprecated('0.5', alternative='`regions.Regions.serialize`')
 def ds9_objects_to_string(regions, coordsys='fk5', fmt='.6f', radunit='deg'):
     """
     Convert a list of `~regions.Region` objects to a DS9 region string.
-
-    See :ref:`gs-ds9`
 
     Parameters
     ----------
@@ -32,16 +44,16 @@ def ds9_objects_to_string(regions, coordsys='fk5', fmt='.6f', radunit='deg'):
     region_string : str
         A DS9 region string.
     """
-    shapelist = _to_shape_list(regions, coordsys)
-    return shapelist.to_ds9(coordsys, fmt, radunit)
+    return _serialize_ds9(regions, coordsys=coordsys, fmt=fmt, radunit=radunit)
 
 
-def write_ds9(regions, filename, coordsys='fk5', fmt='.6f', radunit='deg'):
+@RegionsRegistry.register(Region, 'write', 'ds9')
+@RegionsRegistry.register(Regions, 'write', 'ds9')
+def _write_ds9(regions, filename, coordsys='fk5', fmt='.6f', radunit='deg',
+               overwrite=False):
     """
     Convert a list of `~regions.Region` to a DS9 string and write to a
     file.
-
-    See :ref:`gs-ds9`
 
     Parameters
     ----------
@@ -61,7 +73,49 @@ def write_ds9(regions, filename, coordsys='fk5', fmt='.6f', radunit='deg'):
 
     radunit : str, optional
         The unit of the radius.
+
+    overwrite : bool, optional
+        If True, overwrite the output file if it exists. Raises an
+        `OSError` if False and the output file exists. Default is False.
     """
-    output = ds9_objects_to_string(regions, coordsys, fmt, radunit)
+    if os.path.lexists(filename) and not overwrite:
+        raise OSError(f'{filename} already exists')
+
+    output = _serialize_ds9(regions, coordsys=coordsys, fmt=fmt,
+                            radunit=radunit)
     with open(filename, 'w') as fh:
         fh.write(output)
+
+
+@deprecated('0.5', alternative='`regions.Regions.write`')
+def write_ds9(regions, filename, coordsys='fk5', fmt='.6f', radunit='deg',
+              overwrite=False):
+    """
+    Convert a list of `~regions.Region` to a DS9 string and write to a
+    file.
+
+    Parameters
+    ----------
+    regions : list
+        A list of `~regions.Region` objects.
+
+    filename : str
+        The filename in which the string is to be written.
+
+    coordsys : str, optional
+        An Astropy coordinate system that overrides the coordinate
+        frames of all regions.
+
+    fmt : str, optional
+        A python string format defining the output precision. Default is
+        '.6f', which is accurate to 0.0036 arcseconds.
+
+    radunit : str, optional
+        The unit of the radius.
+
+    overwrite : bool, optional
+        If True, overwrite the output file if it exists. Raises an
+        `OSError` if False and the output file exists. Default is False.
+    """
+    _write_ds9(regions, filename, coordsys=coordsys, fmt=fmt,
+               radunit=radunit, overwrite=overwrite)
